@@ -15,48 +15,48 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.0.3",
-    "@angular/cdk": "^7.0.4",
-    "@angular/common": "^7.0.3",
-    "@angular/compiler": "^7.0.3",
-    "@angular/core": "^7.0.3",
-    "@angular/forms": "^7.0.3",
-    "@angular/http": "^7.0.3",
-    "@angular/material": "^7.0.4",
-    "@angular/platform-browser": "^7.0.3",
-    "@angular/platform-browser-dynamic": "^7.0.3",
-    "@angular/router": "^7.0.3",
-    "@swimlane/ngx-charts": "^10.0.0",
+    "@angular-devkit/build-angular": "^0.803.13",
+    "@angular/animations": "^8.2.11",
+    "@angular/cdk": "^8.2.3",
+    "@angular/common": "^8.2.11",
+    "@angular/compiler": "^8.2.11",
+    "@angular/core": "^8.2.11",
+    "@angular/forms": "^8.2.11",
+    "@angular/http": "^7.2.15",
+    "@angular/material": "^8.2.3",
+    "@angular/platform-browser": "^8.2.11",
+    "@angular/platform-browser-dynamic": "^8.2.11",
+    "@angular/router": "^8.2.11",
+    "@swimlane/ngx-charts": "^12.0.0",
     "bootstrap": ">=4.3.1",
-    "core-js": "^2.5.7",
+    "core-js": "^3.3.3",
     "hammerjs": "^2.0.8",
     "material-design-icons": "^3.0.1",
-    "material-icons": "^0.2.3",
-    "ngx-mqtt": "^6.7.0",
+    "material-icons": "^0.3.1",
+    "ngx-mqtt": "^6.13.2",
     "roboto-fontface": "^0.10.0",
-    "rxjs": "^6.3.3",
-    "rxjs-compat": "^6.0.0-rc.0",
-    "zone.js": "^0.8.26"
+    "rxjs": "6.5.3",
+    "zone.js": "^0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.12.0",
-    "@angular/cli": "^7.0.5",
-    "@angular/compiler-cli": "^7.0.3",
-    "@angular/language-service": "^7.0.3",
-    "@types/jasmine": "2.8.11",
-    "@types/node": "^10.12.7",
-    "codelyzer": "4.5.0",
-    "jasmine-core": "~3.3.0",
+    "@angular/cli": "^8.3.13",
+    "@angular/compiler-cli": "^8.2.11",
+    "@angular/language-service": "^8.2.11",
+    "@types/jasmine": "3.4.4",
+    "@types/node": "^12.11.5",
+    "codelyzer": "5.2.0",
+    "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "^4.2.1",
-    "karma": "^3.1.1",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^2.0.4",
-    "karma-jasmine": "^1.1.2",
-    "karma-jasmine-html-reporter": "^1.4.0",
-    "protractor": "^5.4.1",
-    "ts-node": "~7.0.1",
-    "tslint": "~5.11.0",
-    "typescript": "^3.1.6"
+    "karma": "^4.4.1",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-cli": "~2.0.0",
+    "karma-coverage-istanbul-reporter": "^2.1.0",
+    "karma-jasmine": "^2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.2",
+    "node-sass": "^4.12.0",
+    "protractor": "^5.4.2",
+    "ts-node": "~8.4.1",
+    "tslint": "~5.20.0",
+    "typescript": "3.5.3"
   }
 }
