@@ -1,3 +1,5 @@
 defmodule Sense.Repo do
-  use Ecto.Repo, otp_app: :sense
+  use Ecto.Repo,
+    otp_app: :sense,
+    adapter: Ecto.Adapters.Postgres
 end
