@@ -18,16 +18,6 @@ defmodule Sense.Router do
     plug :accepts, ["json"]
   end
 
-  pipeline :with_session do
-    plug Guardian.Plug.VerifySession
-    plug Guardian.Plug.LoadResource
-    plug Sense.CurrentUser
-  end
-
-  pipeline :session_required do
-    plug Guardian.Plug.EnsureAuthenticated, handler: Sense.GuardianErrorHandler
-  end
-
   scope "/mqtt", Sense do
     post "/user", MqttAuthenticatorController, :user
     post "/superuser", MqttAuthenticatorController, :superuser
@@ -35,15 +25,13 @@ defmodule Sense.Router do
   end
 
   scope "/", Sense do
-    pipe_through [:browser, :with_session]
+    pipe_through [:browser]
 
     get "/", PageController, :index
-    resources "/users", UserController, only: [:new, :create]
-    resources "/sessions", SessionController, only: [:new, :create, :delete]
   end
 
   scope "/", Sense do
-    pipe_through [:browser, :with_session, :session_required]
+    pipe_through [:browser]
 
     resources "/users", UserController, only: [:show]
   end
