@@ -1,66 +1,58 @@
 import { Injectable } from '@angular/core';
-import { Headers, Http } from '@angular/http';
+import { HttpHeaders, HttpClient } from '@angular/common/http';
 import { environment } from '../environments/environment';
 
-import 'rxjs/add/operator/toPromise';
-
 import { Actuator } from './actuator';
+import { catchError, map } from 'rxjs/operators';
+
+import { Observable, of } from 'rxjs';
 
 @Injectable()
 export class ActuatorService {
 
-  private headers = new Headers({'Content-Type': 'application/json'});
+  private headers = new HttpHeaders({'Content-Type': 'application/json'});
   private devicesUrl = environment.apiUrl + '/api/v1/devices';
 
-  constructor(private http: Http) { }
+  constructor(private http: HttpClient) { }
 
-  getActuators(device_id: number): Promise<Actuator[]> {
+  getActuators(device_id: number): Observable<Actuator[]> {
     return this.http.get(`${this.devicesUrl}/${device_id}/actuators`)
-               .toPromise()
-               .then(response => response.json().data as Actuator[])
-               .catch(this.handleError);
+      .pipe(
+        map( (response: any) => response.data as Actuator[])
+      );
   }
 
-
-  getActuator(device_id: number, id: number): Promise<Actuator> {
+  getActuator(device_id: number, id: number): Observable<Actuator> {
     const url = `${this.devicesUrl}/${device_id}/actuators/${id}`;
     return this.http.get(url)
-      .toPromise()
-      .then(response => response.json().data as Actuator)
-      .catch(this.handleError);
+      .pipe(
+        map( (response: any) => response.data as Actuator)
+      );
   }
 
-  delete(actuator: Actuator): Promise<void> {
+  delete(actuator: Actuator): Observable<Object> {
     const url = `${this.devicesUrl}/${actuator.device_id}/actuators/${actuator.id}`;
     return this.http.delete(url, {headers: this.headers})
-      .toPromise()
-      .then(() => null)
-      .catch(this.handleError);
   }
 
-  create(actuator: Actuator): Promise<Actuator> {
+  create(actuator: Actuator): Observable<Actuator> {
     actuator.value = +actuator.value
     return this.http
       .post( `${this.devicesUrl}/${actuator.device_id}/actuators`, JSON.stringify({actuator: actuator}), {headers: this.headers})
-      .toPromise()
-      .then(res => res.json().data as Actuator)
-      .catch(this.handleError);
+      .pipe(
+        map( (response: any) => response.data as Actuator)
+      );
   }
 
-  update(actuator: Actuator): Promise<Actuator> {
+  update(actuator: Actuator): Observable<Actuator> {
     actuator.value = +actuator.value
     const body: any = { actuator: actuator };
     const url = `${this.devicesUrl}/${actuator.device_id}/actuators/${actuator.id}`;
 
     return this.http
       .put(url, JSON.stringify(body), {headers: this.headers})
-      .toPromise()
-      .then(() => actuator)
-      .catch(this.handleError);
-  }
-
-  private handleError(error: any): Promise<any> {
-    console.error('An error occurred', error); // for demo purposes only
-    return Promise.reject(error.message || error);
+      .pipe(
+        map( (response: any) => response.data as Actuator)
+      );
   }
 }
