@@ -1,64 +1,63 @@
 import { Injectable } from '@angular/core';
-import { Headers, Http } from '@angular/http';
+import { HttpHeaders, HttpClient } from '@angular/common/http';
 import { environment } from '../environments/environment';
 
-import 'rxjs/add/operator/toPromise';
-
 import { Device } from './device';
+
+import { Observable, of } from 'rxjs';
+import { catchError, map } from 'rxjs/operators';
 
 @Injectable()
 export class DeviceService {
 
-  private headers = new Headers({'Content-Type': 'application/json'});
+  private headers = new HttpHeaders({'Content-Type': 'application/json'});
   private devicesUrl = environment.apiUrl + '/api/v1/devices';
 
-  constructor(private http: Http) { }
+  constructor(private http: HttpClient) { }
 
-  getDevices(): Promise<Device[]> {
-    return this.http.get(this.devicesUrl)
-               .toPromise()
-               .then(response => response.json().data as Device[])
-               .catch(this.handleError);
-  }
-
-
-  getDevice(id: number): Promise<Device> {
-    const url = `${this.devicesUrl}/${id}`;
-    return this.http.get(url)
-      .toPromise()
-      .then(response => response.json().data as Device)
-      .catch(this.handleError);
-  }
-
-  delete(id: number): Promise<void> {
-    const url = `${this.devicesUrl}/${id}`;
-    return this.http.delete(url, {headers: this.headers})
-      .toPromise()
-      .then(() => null)
-      .catch(this.handleError);
-  }
-
-  create(device: Device): Promise<Device> {
+  getDevices(): Observable<Device[]> {
     return this.http
-      .post(this.devicesUrl, JSON.stringify({device: device}), {headers: this.headers})
-      .toPromise()
-      .then(res => res.json().data as Device)
-      .catch(this.handleError);
+      .get(this.devicesUrl)
+      .pipe(
+        map( (response: any) => response.data as Device[] )
+      );
   }
 
-  update(device: Device): Promise<Device> {
+
+  getDevice(id: number): Observable<Device> {
+    const url = `${this.devicesUrl}/${id}`;
+    return this.http
+      .get(url)
+      .pipe(
+        map( (response: any) => response.data as Device )
+      );
+  }
+
+  delete(id: number): Observable<Object> {
+    const url = `${this.devicesUrl}/${id}`;
+    return this.http
+      .delete(url, {headers: this.headers})
+  }
+
+  create(device: Device): Observable<Device> {
+    const body: any = { device: device };
+    return this.http
+      .post(this.devicesUrl,
+            JSON.stringify(body),
+            {headers: this.headers})
+      .pipe(
+        map( (response: any) => response as Device )
+      );
+  }
+
+  update(device: Device): Observable<Device> {
     const body: any = { device: device };
     const url = `${this.devicesUrl}/${device.id}`;
 
     return this.http
       .put(url, JSON.stringify(body), {headers: this.headers})
-      .toPromise()
-      .then(() => device)
-      .catch(this.handleError);
-  }
-
-  private handleError(error: any): Promise<any> {
-    console.error('An error occurred', error); // for demo purposes only
-    return Promise.reject(error.message || error);
+      .pipe(
+        map( (response: any) => response as Device )
+      );
   }
 }
