@@ -1,4 +1,3 @@
-import 'rxjs/add/operator/switchMap';
 import { Component, OnInit, Input } from '@angular/core';
 import { ActivatedRoute, Params } from '@angular/router';
 import { Location } from '@angular/common';
@@ -32,7 +31,7 @@ export class MeasureComponent implements OnInit {
 
   create(): void {
     this.measureService.create(this.measure)
-      .then(() => this.goBack());
+      .subscribe(() => this.goBack());
   }
 
   goBack(): void {
