@@ -1,8 +1,8 @@
 export const environment = {
   production: true,
-  serverUrl: 'http://senseapp.space',
-  apiUrl: 'http://api.senseapp.space',
-  mqttHost: 'mqtt.senseapp.space',
-  mqttPort: 8083,
-  mqttPath: '/mqtt'
+  serverUrl: 'https://senseapp.space',
+  apiUrl: 'https://api.senseapp.space',
+  mqttHost: 'broker.senseapp.space',
+  mqttPort: 443,
+  mqttPath: '/wss'
 };
