@@ -42,7 +42,7 @@
 
 
 /** Evergreen browsers require these. **/
-import 'core-js/es6/reflect';
+import 'core-js/es/reflect';
 
 
 
